+++ DearPyGui/src/core/AppItems/composite/mvStyleWindow.cpp
@@ -1,6 +1,6 @@
 #include <imgui.h>
 #include "mvApp.h"
-#include "mvAppItems.h"
+#include "mvAppItemCommons.h"
 #include "mvStyleWindow.h"
 #include "mvInput.h"
 #include "mvItemRegistry.h"
