+++ DearPyGui/src/core/AppItems/mvItemRegistry.cpp
@@ -2,7 +2,7 @@
 #include "mvProfiler.h"
 #include "mvApp.h"
 #include "mvItemRegistry.h"
-#include "mvAppItems.h"
+#include "mvAppItemCommons.h"
 #include "mvLog.h"
 
 namespace Marvel {
