+++ DearPyGui/src/core/Modules/mvModule_Core.cpp
@@ -2,7 +2,7 @@
 #include "mvApp.h"
 #include "mvInput.h"
 #include "mvAppLog.h"
-#include "mvAppItems.h"
+#include "mvAppItemCommons.h"
 #include "mvWindow.h"
 #include "mvFontManager.h"
 #include "mvItemRegistry.h"
