+++ DearPyGui/src/core/Theming/mvThemeManager.cpp
@@ -1,6 +1,6 @@
 #include "mvThemeManager.h"
 #include "imgui.h"
-#include "mvAppItems.h"
+#include "mvAppItemCommons.h"
 #include "mvCore.h"
 #include "mvItemRegistry.h"
